Add irqbalanceBanned cores to PerformanceProfile API
PerformanceProfile API gains a field IrqbalanceBanned which defines a
set of CPUs that will not be used by irqbalance when balancing
interrupts across CPUs. By default, irqbalance will consider all CPUs
for load balancing except for CPUs which are assigned to pods with
annotation "irq-load-balancing.crio.io: true". This option allows to
expand this list of dynamically banned CPUs with a statically defined
set of CPUs.